Change zoom level to power of 2
Default zoom is 2 blocks per pixel,  zoom in is a max of 4 pixels per block, zoom out is 1 chunk per pixel. 

This makes exporting images easier as each pixel could be exactly 1 block per pixel or 1 chunk per 1 pixel.

If zoom multiply/divide by 2 is too much, one could do sqrt(2) or cubic root of 2.

Also getUnitsToScroll is based on the windows setting, if I set windows settings to 2, minemap doesn't zoom, if the setting is 1, it's inverted, and if it's 100, then I'm stuck at either the maxZoom or minZoom.